Quando desativar um abrigo também desativar os pets. #76
close #76
