+++ adopet/accounts/tests/shelter/test_delete.py
@@ -30,6 +30,28 @@ def test_positive_by_id(client_api_auth_shelter, shelter):
     assert not shelter.is_active
 
 
+def test_positive_must_disable_the_pets_also(client_api_auth_shelter, shelter, pets):
+    """Must also disable the pets in this shelter"""
+
+    pk = shelter.pk
+
+    url = resolve_url(URL, pk=pk)
+
+    resp = client_api_auth_shelter.delete(url)
+
+    assert resp.status_code == status.HTTP_200_OK
+
+    shelter.refresh_from_db()
+    db_pets = shelter.pets.all()
+
+    body = resp.json()
+
+    assert body["msg"] == "Abrigo deletado com sucesso."
+    assert not shelter.is_active
+    for p in db_pets:
+        assert not p.is_active
+
+
 def test_negative_only_own_shelter_can_delete_itself(client_api_auth_shelter, shelter, users):
     pk = User.objects.shelter().exclude(id=shelter.pk).values_list("pk").first()[0]
